pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ragbench"
version = "0.1.0"
description = "Large-Scale Retrieval Evaluation & Optimization System for RAG"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Gokulakrishnan Thiruvikraman"}]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "rank-bm25>=0.2.2",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2",
    "tiktoken>=0.5",
    "mlflow>=2.9",
    "optuna>=3.4",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.7",
]
graph = ["neo4j>=5.0"]
pinecone = ["pinecone-client>=3.0"]
all = ["ragbench[dev,graph,pinecone]"]

[project.scripts]
ragbench = "ragbench.cli:main"

[tool.setuptools.packages.find]
include = ["ragbench*"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
